Time: 0 ms (100.00%) | Memory: 7.5 MB (69.46%) - LeetSync

+++ 62-unique-paths/unique-paths.cpp
@@ -3,7 +3,7 @@ public:
     int uniquePaths(int m, int n){
         int N = m + n - 2;    
         int r = m - 1;
-        double ans = 1;
+        long long ans = 1;
         for(int i = 1;i <= r;i++){
             ans = ans * (N - r + i)/i;
         }
